Rust: Move example to separate module
